% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast}
\alias{forecast}
\title{Forecast function
Forecasts for the data}
\usage{
forecast(x, sm, sy, em, ey, fort)
}
\arguments{
\item{x}{data vector to analyse}

\item{sm}{start month}

\item{sy}{start year}

\item{em}{end month}

\item{ey}{end year}

\item{fort}{forecast length in months}
}
\description{
Forecast function
Forecasts for the data
}
\examples{
forecast(x, sm, sy, em, ey, fort)
}
